+++ src/main/java/org/dma/gbdt4spark/objective/loss/BinaryLoss.java
@@ -1,9 +1,9 @@
 package org.dma.gbdt4spark.objective.loss;
 
 public interface BinaryLoss extends Loss {
-    float firOrderGrad(float pred, float label);
+    double firOrderGrad(float pred, float label);
 
-    float secOrderGrad(float pred, float label);
+    double secOrderGrad(float pred, float label);
 
-    float secOrderGrad(float pred, float label, float firGrad);
+    double secOrderGrad(float pred, float label, double firGrad);
 }
